split out scaled channel header
